start adding cross-docking analysis code
