+++ harbor/analysis.py
@@ -1,21 +1,8 @@
+from harbor.analysis.utils import get_ci_from_bootstraps
 from harbor.schema.data import ActiveInactiveDataset
-from harbor.plot_schema import RocCurve, PrecisionRecallCurve, RocCurveUncertainty
-from sklearn.metrics import roc_curve, roc_auc_score, precision_recall_curve, auc
-from pydantic import BaseModel, Field
+from harbor.plot_schema import RocCurve, RocCurveUncertainty
+from sklearn.metrics import roc_curve, roc_auc_score
 import numpy as np
-
-
-def get_ci_from_bootstraps(
-    values: list[float], alpha: float = 0.95
-) -> tuple[float, float]:
-    """
-    Calculate the confidence interval of a list of values
-    """
-    sorted_values = np.sort(values)
-    n_values = len(sorted_values)
-    lower_index = sorted_values[int(n_values * (1 - alpha))]
-    upper_index = sorted_values[int(n_values * alpha)]
-    return (lower_index, upper_index)
 
 
 def get_roc_curve(dataset: ActiveInactiveDataset, model_id: str) -> RocCurve:
